server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xlsxworld-backend"
version = "0.1.0"
description = "XLSX World backend API"
requires-python = ">=3.13"
license = "MIT"
keywords = ["xlsx", "fastapi", "backend"]
classifiers = [
  "Framework :: FastAPI",
  "Programming Language :: Python :: 3.13",
]

dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.10.0",
  "certifi==2025.8.3",
  "click==8.2.1",
  "colorama==0.4.6",
  "dnspython==2.8.0",
  "email-validator==2.3.0",
  "et_xmlfile==2.0.0",
  "fastapi==0.116.1",
  "fastapi-cli==0.0.11",
  "fastapi-cloud-cli==0.1.5",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.6.4",
  "httpx==0.28.1",
  "idna==3.10",
  "Jinja2==3.1.6",
  "markdown-it-py==4.0.0",
  "MarkupSafe==3.0.2",
  "mdurl==0.1.2",
  "openpyxl==3.1.5",
  "pyxlsb==1.0.10",
  "pydantic==2.12.5",
  "pydantic_core==2.41.5",
  "Pygments==2.19.2",
  "python-dotenv==1.1.1",
  "python-multipart==0.0.20",
  "PyYAML==6.0.2",
  "rich==14.1.0",
  "rich-toolkit==0.15.1",
  "rignore==0.6.4",
  "sentry-sdk==2.37.1",
  "shellingham==1.5.4",
  "sniffio==1.3.1",
  "starlette==0.47.3",
  "typer==0.17.4",
  "typing-inspection==0.4.2",
  "typing_extensions==4.15.0",
  "urllib3==2.5.0",
  "uvicorn==0.39.0",
  "uv>=0.11.1",
  "watchfiles==1.1.0",
  "websockets==15.0.1",
  "xlrd==2.0.1",
  "psycopg[binary]==3.2.13",
  "sqlalchemy==2.0.36",
  "passlib==1.7.4",
  "python-jose==3.3.0",
  "bcrypt==4.2.1",
  "slowapi>=0.1.9",
]

[tool.setuptools.packages.find]
where = ["."]
